UPD Extract tournament tree generation
