+++ src/pages/createTournament/CreateTournament.jsx
@@ -1,5 +1,6 @@
 import { useState } from "react";
 import { Form, redirect } from "react-router-dom";
+import { cloneDeep } from "lodash";
 
 import { findDuplicates, random } from "../../utils.js";
 
@@ -49,6 +50,31 @@ const actionCreate = async ({ request, _params }) => {
       tournament.teams[counter].players = playersForTeam;
     }
   }
+
+  // Duels tree generation
+  const teamsBuffer = cloneDeep(tournament.teams);
+  const brackets = [];
+  const round0 = [];
+  while (teamsBuffer.length) {
+    const match = { teams: [], result: [0, 0], matchDone: false };
+    const firstIndex = random(teamsBuffer.length);
+    match.teams.push(teamsBuffer[firstIndex]);
+    teamsBuffer.splice(firstIndex, 1);
+    const secondIndex = random(teamsBuffer.length);
+    match.teams.push(teamsBuffer[secondIndex]);
+    teamsBuffer.splice(secondIndex, 1);
+    round0.push(match);
+  }
+  brackets.push(round0);
+  for (let counter = tournament.teams.length / 2; counter > 1; counter /= 2) {
+    const round = [];
+    for (let bracket = 0; bracket < counter / 2; bracket++) {
+      round.push({ teams: [], result: [0, 0], matchDone: false });
+    }
+    brackets.push(round);
+  }
+
+  tournament.brackets = brackets;
 
   tournament = encodeURI(JSON.stringify(tournament));
   return redirect(`/tournament?tournament=${tournament}`);
